Added User Login Checks.

+++ app/code/Ecommistry/Blog/Controller/Index/Delete.php
@@ -2,12 +2,12 @@
 
 namespace Ecommistry\Blog\Controller\Index;
 
-use Ecommistry\Blog\Model\Blog;
 use Ecommistry\Blog\Model\BlogFactory;
 use Ecommistry\Blog\Model\ResourceModel\BlogFactory as BlogResourceFactory;
+use Magento\Customer\Model\Session;
 use Magento\Framework\App\Action\Action;
 use Magento\Framework\App\Action\Context;
-use Magento\Framework\App\ResponseInterface;
+use Magento\Framework\Controller\ResultFactory;
 
 /**
  * Delete Controller
@@ -25,39 +25,52 @@ use Magento\Framework\App\ResponseInterface;
  */
 class Delete extends Action
 {
+    /** @var \Ecommistry\Blog\Model\BlogFactory */
     private $blogFactory;
+    /** @var \Ecommistry\Blog\Model\ResourceModel\BlogFactory */
     private $blogResourceFactory;
+    /** @var \Magento\Customer\Model\Session */
+    private $session;
     
     public function __construct(
         Context $context,
         BlogFactory $blogFactory,
-        BlogResourceFactory $blogResourceFactory
+        BlogResourceFactory $blogResourceFactory,
+        Session $session
     ) {
+        $this->session = $session;
         $this->blogFactory = $blogFactory;
         $this->blogResourceFactory = $blogResourceFactory;
         parent::__construct($context);
     }
     
     /**
-     * Execute action based on request and return result
-     *
-     * Note: Request will be added as operation argument in future
-     *
-     * @return \Magento\Framework\Controller\ResultInterface|ResponseInterface
-     * @throws \Magento\Framework\Exception\NotFoundException
+     * @return \Magento\Framework\Controller\ResultInterface
      */
     public function execute()
     {
-        $blogId = $this->getRequest()->getParam('id');
-        try {
-            $blog = $this->blogFactory->create();
-            $this->blogResourceFactory->create()
-                ->load($blog, $blogId, 'blog_id');
-            $this->blogResourceFactory->create()->delete($blog);
-            $this->messageManager->addSuccessMessage('Blog Post Deleted');
-        } catch (\Exception $e) {
-            $this->messageManager->addErrorMessage($e->getMessage());
+        if (!$this->session->isLoggedIn()) {
+            $result
+                = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
+            $result->setPath('*/*/message');
+        } else {
+            $result
+                = $this->resultFactory->create(ResultFactory::TYPE_REDIRECT);
+            $result->setUrl('/blog');
+        
+            $blogId = $this->getRequest()->getParam('id');
+            if ($blogId) {
+                try {
+                    $blog = $this->blogFactory->create();
+                    $this->blogResourceFactory->create()
+                        ->load($blog, $blogId, 'blog_id');
+                    $this->blogResourceFactory->create()->delete($blog);
+                    $this->messageManager->addSuccessMessage('Blog Post Deleted');
+                } catch (\Exception $e) {
+                    $this->messageManager->addErrorMessage($e->getMessage());
+                }
+            }
         }
-        $this->_redirect('blog');
+        return $result;
     }
 }
